+++ src/frontend/apps/impress/src/features/docs/docs-grid/components/DocsGridItem.tsx
@@ -9,7 +9,7 @@ import { useConfig } from '@/core';
 import { useCunninghamTheme } from '@/cunningham';
 import { Doc, LinkReach, SimpleDocItem } from '@/docs/doc-management';
 import { DocShareModal } from '@/docs/doc-share';
-import { useDate } from '@/hook';
+import { useDate } from '@/hooks';
 import { useResponsiveStore } from '@/stores';
 
 import { useResponsiveDocGrid } from '../hooks/useResponsiveDocGrid';
