--- src/frontend/apps/impress/src/hook/index.ts
@@ -1,2 +1,0 @@
-export * from './useDate';
-export * from './useClipboard';
