+++ src/frontend/apps/impress/src/hooks/index.ts
@@ -1,1 +1,4 @@
+export * from './useClipboard';
+export * from './useCmdK';
+export * from './useDate';
 export * from './useKeyboardAction';
